pyproject.toml
[tool.poetry]
name = "iot-monitoring"
version = "0.1.0"
description = "A GitHub template repository for a Python"
authors = ["ks6088ts <ks6088ts@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.3"
python-dotenv = "^1.0.1"
azure-iot-device = "^2.13.1"
azure-storage-blob = "^12.20.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.0"
black = "^24.4.0"
pytest = "^8.1.1"
coverage = "^7.4.4"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["D203"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.pytest.ini_options]
addopts = "-ra --cov"
testpaths = ["tests"]
pythonpath = ['.']

[tool.coverage.report]
show_missing = true
